refactor(tray): merge TrayIconUpdater into TrayIconManager
Removed TrayIconUpdater and moved its logic into TrayIconManager, which now handles tray icon updates directly based on sync status. Service registration for TrayIconUpdater was also removed.
