+++ src/gui4/windows/kDrive client/kDrive client/kDrive client.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <AssemblyVersion>3.8.2.1</AssemblyVersion>
     <OutputType>WinExe</OutputType>
@@ -209,7 +209,8 @@
       <Generator>PublicResXFileCodeGenerator</Generator>
     </PRIResource>
     <PRIResource Update="Strings\fr-FR\Resources.resw">
-      <Generator></Generator>
+      <Generator>
+      </Generator>
     </PRIResource>
     <PRIResource Update="Strings\it-IT\Resources.resw">
       <SubType>Designer</SubType>
@@ -289,7 +290,7 @@
     <ItemGroup>
       <!-- Keep only fr-FR, en-US, es-ES, it-IT, de-DE -->
       <RemovingFiles Include="$(OutDir)*\*.mui" Exclude="&#xD;&#xA;            $(OutDir)fr-FR\*.mui;&#xD;&#xA;            $(OutDir)en-US\*.mui;&#xD;&#xA;            $(OutDir)es-ES\*.mui;&#xD;&#xA;            $(OutDir)it-IT\*.mui;&#xD;&#xA;            $(OutDir)de-DE\*.mui" />
-      <RemovingFolders Include="@(RemovingFiles->'%(RootDir)%(Directory)')" />
+      <RemovingFolders Include="@(RemovingFiles-&gt;'%(RootDir)%(Directory)')" />
     </ItemGroup>
     <RemoveDir Directories="@(RemovingFolders)" />
   </Target>
